+++ konva.js
@@ -3,7 +3,7 @@
  * Konva JavaScript Framework v1.2.2
  * http://konvajs.github.io/
  * Licensed under the MIT or GPL Version 2 licenses.
- * Date: Wed Dec 07 2016
+ * Date: Thu Dec 08 2016
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - 2015 by Anton Lavrenov (Konva)
@@ -12918,6 +12918,7 @@
     var AUTO = 'auto',
         //CANVAS = 'canvas',
         CENTER = 'center',
+        JUSTIFY = 'justify',
         CHANGE_KONVA = 'Change.konva',
         CONTEXT_2D = '2d',
         DASH = '-',
@@ -13083,6 +13084,7 @@
                 lineHeightPx = this.getLineHeight() * textHeight,
                 textArr = this.textArr,
                 textArrLen = textArr.length,
+                align = this.getAlign(),
                 totalWidth = this.getWidth(),
                 letterSpacing = this.getLetterSpacing(),
                 textDecoration = this.textDecoration(),
@@ -13112,10 +13114,10 @@
 
                 // horizontal alignment
                 context.save();
-                if(this.getAlign() === RIGHT) {
+                if(align === RIGHT) {
                     context.translate(totalWidth - width - p * 2, 0);
                 }
-                else if(this.getAlign() === CENTER) {
+                else if(align === CENTER) {
                     context.translate((totalWidth - width - p * 2) / 2, 0);
                 }
 
@@ -13131,9 +13133,15 @@
                   context.stroke();
                   context.restore();
                 }
-                if (letterSpacing !== 0) {
+                if ((letterSpacing !== 0) || align === JUSTIFY) {
+                //   var words = text.split(' ');
+                  var spacesNumber = text.split(' ').length - 1;
                   for(var li = 0; li < text.length; li++) {
                     var letter = text[li];
+                    // skip justify for the last line
+                    if ((letter === ' ') && (n !== textArrLen - 1)) {
+                        context.translate(Math.ceil((totalWidth - width) / spacesNumber), 0);
+                    }
                     this.partialText = letter;
                     context.fillStrokeShape(this);
                     context.translate(Math.round(this._getTextSize(letter).width) + letterSpacing, 0);
@@ -13445,7 +13453,7 @@
     Konva.Factory.addGetterSetter(Konva.Text, 'align', LEFT);
 
     /**
-     * get/set horizontal align of text.  Can be 'left', 'center', or 'right'
+     * get/set horizontal align of text.  Can be 'left', 'center', 'right' or 'justify'
      * @name align
      * @method
      * @memberof Konva.Text.prototype
@@ -15241,6 +15249,7 @@
             var that = this;
             var size = this._getTextSize(this.attrs.text);
             var letterSpacing = this.getLetterSpacing();
+            var align = this.align();
 
             this.textWidth = size.width;
             this.textHeight = size.height;
@@ -15258,14 +15267,15 @@
             }
 
             var offset = 0;
-            if (this.align() === 'center') {
+            if (align === 'center') {
               offset = Math.max(0, fullPathWidth / 2 - textFullWidth / 2);
             }
-            if (this.align() === 'right') {
+            if (align === 'right') {
               offset = Math.max(0, fullPathWidth - textFullWidth);
             }
 
             var charArr = this.getText().split('');
+            var spacesNumber = this.getText().split(' ').length - 1;
 
             var p0, p1, pathCmd;
 
@@ -15306,6 +15316,10 @@
             var findSegmentToFitCharacter = function(c) {
 
                 var glyphWidth = that._getTextSize(c).width + letterSpacing;
+
+                if ((c === ' ') && (align === 'justify')) {
+                    glyphWidth += (fullPathWidth - textFullWidth) / spacesNumber;
+                }
 
                 var currLen = 0;
                 var attempts = 0;
@@ -15545,7 +15559,7 @@
      Konva.Factory.addGetterSetter(Konva.TextPath, 'align', 'left');
 
      /**
-      * get/set horizontal align of text.  Can be 'left', 'center', or 'right'
+      * get/set horizontal align of text.  Can be 'left', 'center', 'right' or 'justify'
       * @name align
       * @method
       * @memberof Konva.Text.prototype